Enable main app call in root
