+++ src/game_phases.py
@@ -8,15 +8,10 @@ from src.components.hand import Hand
 from src.components.hand_side import HandSide
 from src.components.player import Player
 from src.components.scoreboard import Scoreboard
-from src.config import Config
 from src.global_state import GlobalState
 from src.services.music_service import MusicService
 from src.services.visualization_service import VisualizationService
 from src.utils.tools import update_background_using_scroll, update_press_key, is_close_app_event
-
-FramePerSec = pygame.time.Clock()
-
-pygame.init()
 
 GlobalState.load_main_screen()
 VisualizationService.load_main_game_displays()
@@ -38,22 +33,23 @@ all_sprites.add(H1)
 all_sprites.add(H2)
 
 
-def game_over():
-    P1.reset()
-    H1.reset()
-    H2.reset()
-    GlobalState.GAME_STATE = GameStatus.MAIN_MENU
-    time.sleep(0.5)
+def main_menu_phase():
+    scoreboard.reset_current_score()
 
+    events = pygame.event.get()
 
-def update_game_display():
-    pygame.display.update()
-    FramePerSec.tick(Config.FPS)
+    for event in events:
+        if is_close_app_event(event):
+            GlobalState.GAME_STATE = GameStatus.GAME_END
+            return
 
+        if event.type == pygame.KEYDOWN:
+            GlobalState.GAME_STATE = GameStatus.GAMEPLAY
 
-def close_app():
-    pygame.quit()
-    sys.exit()
+    GlobalState.SCROLL = update_background_using_scroll(GlobalState.SCROLL)
+    VisualizationService.draw_background_with_scroll(GlobalState.SCREEN, GlobalState.SCROLL)
+    GlobalState.PRESS_Y = update_press_key(GlobalState.PRESS_Y)
+    VisualizationService.draw_main_menu(GlobalState.SCREEN, scoreboard.get_max_score(), GlobalState.PRESS_Y)
 
 
 def gameplay_phase():
@@ -83,36 +79,14 @@ def gameplay_phase():
         game_over()
 
 
-def main_menu_phase():
-    events = pygame.event.get()
-
-    for event in events:
-        if is_close_app_event(event):
-            GlobalState.GAME_STATE = GameStatus.GAME_END
-            return
-
-        if event.type == pygame.KEYDOWN:
-            GlobalState.GAME_STATE = GameStatus.GAMEPLAY
-
-    GlobalState.SCROLL = update_background_using_scroll(GlobalState.SCROLL)
-    VisualizationService.draw_background_with_scroll(GlobalState.SCREEN, GlobalState.SCROLL)
-    GlobalState.PRESS_Y = update_press_key(GlobalState.PRESS_Y)
-    VisualizationService.draw_main_menu(GlobalState.SCREEN, scoreboard.get_max_score(), GlobalState.PRESS_Y)
+def exit_game_phase():
+    pygame.quit()
+    sys.exit()
 
 
-def main():
-    while True:
-        if GlobalState.GAME_STATE == GameStatus.MAIN_MENU:
-            scoreboard.reset_current_score()
-            main_menu_phase()
-        elif GlobalState.GAME_STATE == GameStatus.GAMEPLAY:
-            gameplay_phase()
-        elif GlobalState.GAME_STATE == GameStatus.GAME_END:
-            close_app()
-
-        MusicService.start_background_music()
-        update_game_display()
-
-
-if __name__ == "__main__":
-    main()
+def game_over():
+    P1.reset()
+    H1.reset()
+    H2.reset()
+    GlobalState.GAME_STATE = GameStatus.MAIN_MENU
+    time.sleep(0.5)
